refactor(data): add year progress repository
