+++ app/src/main/java/com/example/sioribi/data/YearProgressRepository.kt
@@ -1,0 +1,7 @@
+package com.example.sioribi.data
+
+import com.example.sioribi.domain.YearProgressModel
+
+interface YearProgressRepository {
+    fun getYearProgress(): YearProgressModel
+}
