+++ app/src/main/java/com/example/sioribi/di/AppGraph.kt
@@ -1,8 +1,10 @@
 package com.example.sioribi.di
 
 import android.content.Context
+import com.example.sioribi.data.DefaultYearProgressRepository
 import com.example.sioribi.data.SystemTimeDataSource
 import com.example.sioribi.data.TimeDataSource
+import com.example.sioribi.data.YearProgressRepository
 import com.example.sioribi.domain.GetYearProgressUseCase
 import com.example.sioribi.ui.WidgetRefreshCoordinator
 import com.example.sioribi.ui.WorkManagerWidgetRefreshEnqueuer
@@ -11,7 +13,10 @@ class AppGraph(
     appContext: Context,
 ) {
     val timeDataSource: TimeDataSource = SystemTimeDataSource()
-    val getYearProgressUseCase: GetYearProgressUseCase = GetYearProgressUseCase(timeDataSource)
+    val yearProgressRepository: YearProgressRepository =
+        DefaultYearProgressRepository(timeDataSource)
+    val getYearProgressUseCase: GetYearProgressUseCase =
+        GetYearProgressUseCase(yearProgressRepository)
     val widgetRefreshCoordinator: WidgetRefreshCoordinator =
         WidgetRefreshCoordinator(WorkManagerWidgetRefreshEnqueuer(appContext))
 }
