+++ app/src/main/java/com/example/sioribi/domain/GetYearProgressUseCase.kt
@@ -1,31 +1,9 @@
 package com.example.sioribi.domain
 
-import com.example.sioribi.data.TimeDataSource
-import kotlin.math.roundToInt
+import com.example.sioribi.data.YearProgressRepository
 
 class GetYearProgressUseCase(
-    private val timeDataSource: TimeDataSource,
+    private val yearProgressRepository: YearProgressRepository,
 ) {
-    companion object {
-        private const val COMMON_YEAR_DAYS = 365
-        private const val LEAP_YEAR_DAYS = 366
-        private const val PERCENT_SCALE = 100
-    }
-
-    fun execute(): YearProgressModel {
-        val today = timeDataSource.today()
-        val year = today.year
-        val totalDays = if (today.isLeapYear) LEAP_YEAR_DAYS else COMMON_YEAR_DAYS
-        val currentDay = today.dayOfYear
-        val progressPercentage =
-            (currentDay.toDouble() / totalDays.toDouble() * PERCENT_SCALE).roundToInt()
-        val formattedString = "$currentDay/$totalDays"
-        return YearProgressModel(
-            currentDay = currentDay,
-            totalDays = totalDays,
-            year = year,
-            progressPercentage = progressPercentage,
-            formattedString = formattedString,
-        )
-    }
+    fun execute(): YearProgressModel = yearProgressRepository.getYearProgress()
 }
